# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Project Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  name: string;
  email: string;
  role: string;
  avatarColor: string;
  createdAt: string;
}

export interface MemberInput {
  name: string;
  email: string;
  role: string;
  avatarColor?: string;
}

export interface MemberUpdate {
  name?: string;
  email?: string;
  role?: string;
  avatarColor?: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  status?: ProjectInputStatus;
  dueDate?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  /** @nullable */
  dueDate?: string | null;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  projectId: number;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  assigneeColor?: string | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  projectId: number;
  assigneeId?: number;
  dueDate?: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  projectId?: number;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  totalTasks: number;
  openTasks: number;
  overdueTasks: number;
  completedTasks: number;
  totalMembers: number;
}

export interface ActivityItem {
  id: number;
  taskId: number;
  taskTitle: string;
  action: string;
  memberName: string;
  memberColor: string;
  projectName: string;
  timestamp: string;
}

export interface ProjectProgress {
  projectId: number;
  projectName: string;
  status: string;
  totalTasks: number;
  doneTasks: number;
  inProgressTasks: number;
  todoTasks: number;
  progress: number;
}

export type ListTasksParams = {
status?: string;
assigneeId?: number;
projectId?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};