# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project operations
  - name: tasks
    description: Task operations
  - name: members
    description: Team member operations
  - name: dashboard
    description: Dashboard and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all team members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  /projects/{id}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks (optionally filtered)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: integer
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get high-level dashboard stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent task activity feed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/progress:
    get:
      operationId: getProjectProgress
      tags: [dashboard]
      summary: Get per-project progress breakdown
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectProgress"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Member:
      type: object
      required: [id, name, email, role, avatarColor, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string

    MemberInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarColor:
          type: string

    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarColor:
          type: string

    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        dueDate:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        dueDate:
          type: ["string", "null"]

    Task:
      type: object
      required: [id, title, status, priority, projectId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        assigneeColor:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string

    TaskInput:
      type: object
      required: [title, projectId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assigneeId:
          type: integer
        dueDate:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]

    DashboardSummary:
      type: object
      required: [totalProjects, activeProjects, totalTasks, openTasks, overdueTasks, completedTasks, totalMembers]
      properties:
        totalProjects:
          type: integer
        activeProjects:
          type: integer
        totalTasks:
          type: integer
        openTasks:
          type: integer
        overdueTasks:
          type: integer
        completedTasks:
          type: integer
        totalMembers:
          type: integer

    ActivityItem:
      type: object
      required: [id, taskId, taskTitle, action, memberName, memberColor, projectName, timestamp]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        taskTitle:
          type: string
        action:
          type: string
        memberName:
          type: string
        memberColor:
          type: string
        projectName:
          type: string
        timestamp:
          type: string

    ProjectProgress:
      type: object
      required: [projectId, projectName, status, totalTasks, doneTasks, inProgressTasks, todoTasks, progress]
      properties:
        projectId:
          type: integer
        projectName:
          type: string
        status:
          type: string
        totalTasks:
          type: integer
        doneTasks:
          type: integer
        inProgressTasks:
          type: integer
        todoTasks:
          type: integer
        progress:
          type: number